+++ src/utils/Utils.ts
@@ -1,5 +1,4 @@
-import { EnumerateRequest } from "../types";
-
+import { EnumerateRequest } from '../types';
 
 /*
   StringUtils is a utility class for string operations.
@@ -10,13 +9,22 @@ class Utils {
    * @param {string} str - The string to encode.
    * @return {string} The encoded string.
    */
-  static createUrlParams(request?: EnumerateRequest) {
+  static createUrlParams(request?: EnumerateRequest): string {
     const params = new URLSearchParams();
     if (request?.token) {
       params.set('token', request.token);
     }
     if (request?.maxKeys) {
       params.set('maxKeys', request.maxKeys.toString());
+    }
+    if (request?.includeData) {
+      params.set('incldata', request.includeData.toString());
+    }
+    if (request?.includeSubordinates) {
+      params.set('inclsub', request.includeSubordinates.toString());
+    }
+    if (request?.skip) {
+      params.set('skip', request.skip.toString());
     }
     return params.toString() ? '?' + params.toString() : '';
   }